back-end code organizing
